+++ src/TMC_2209/TMC_2209_StepperDriver.py
@@ -15,8 +15,9 @@ import math
 import statistics
 import threading
 from RPi import GPIO
-from .TMC_2209_uart import TMC_UART
-from . import TMC_2209_reg as reg
+from .TMC_2209_uart import TMC_UART as tmc_uart
+from . import TMC_2209_reg as tmc_reg
+from . import TMC_2209_math as tmc_math
 
 
 
@@ -82,7 +83,8 @@ class TMC_2209:
 
 
     _msres = -1
-    _steps_per_revolution = 0
+    _steps_per_rev = 0
+    _fullsteps_per_rev = 200
 
     _loglevel = Loglevel.INFO
     _logprefix = "TMC2209"
@@ -91,7 +93,7 @@ class TMC_2209:
     _target_pos = 0                  # the target position in steps
     _speed = 0.0                    # the current speed in steps per second
     _max_speed = 1.0                 # the maximum speed in steps per second
-    _max_speed_homing = 500           # the maximum speed in steps per second for homing
+    _max_speed_homing = 200           # the maximum speed in steps per second for homing
     _acceleration = 1.0             # the acceleration in steps per second per second
     _acceleration_homing = 10000     # the acceleration in steps per second per second for homing
     _sqrt_twoa = 1.0                # Precomputed sqrt(2*_acceleration)
@@ -134,8 +136,19 @@ class TMC_2209:
     driver_address=0, no_uart=False, gpio_mode=GPIO.BCM, loglevel = None):
         """
         init function
+        
+            Parameters:
+                pin_en (int): Pin number of EN pin
+                pin_step (int): Pin number of STEP pin
+                pin_dir (int): Pin number of DIR pin
+                baudrate (int): baudrate exp: 9600, 115200
+                serialport (string): win: 'COM1'; linux: 'dev/serial0'
+                driver_address (int): 0 - 3
+                no_uart=False (bool): skip UART init, if only STEP/DIR is used
+                gpio_mode (enum): numbering system for the GPIO pins
+                loglevel (enum): level for which to log
         """
-        self.tmc_uart = TMC_UART(serialport, baudrate, driver_address)
+        self.tmc_uart = tmc_uart(serialport, baudrate, driver_address)
 
         if loglevel is not None:
             self._loglevel = loglevel
@@ -161,7 +174,7 @@ class TMC_2209:
         self.log("GPIO Init finished", Loglevel.INFO.value)
 
         if not no_uart:
-            self.read_steps_per_revolution()
+            self.read_steps_per_rev()
             self.clearGSTAT()
 
         self.tmc_uart.flush_serial_buffer()
@@ -207,6 +220,9 @@ class TMC_2209:
     def set_loglevel(self, loglevel):
         """
         set the loglevel. See the Enum Loglevel
+
+            Parameters:
+                loglevel (enum): level for which to log
         """
         self._loglevel = loglevel
 
@@ -215,6 +231,10 @@ class TMC_2209:
     def log(self, message, loglevel=Loglevel.NONE.value):
         """
         logs a message
+
+            Parameters:
+                message (string): message to log
+                loglevel (enum): level for which to log
         """
         if self._loglevel.value >= loglevel:
             print(self._logprefix+"_"+str(self.tmc_uart.mtr_id)+": "+message)
@@ -225,6 +245,9 @@ class TMC_2209:
         """
         set whether the movment should be relative or absolute by default.
         See the Enum MovementAbsoluteRelative
+
+            Paramters:
+                movement_abs_rel (enum): whether the movment should be relative or absolute
         """
         self._movement_abs_rel = movement_abs_rel
 
@@ -232,50 +255,54 @@ class TMC_2209:
 
     def readDRVSTATUS(self):
         """
-        read the register Adress "DRVSTATUS" and prints all current setting
+        read the register Adress "DRV_STATUS" and prints all current setting
+
+            Returns:
+                drvstatus (int): 32bit DRV_STATUS Register
+
         """
         self.log("---")
         self.log("DRIVER STATUS:")
-        drvstatus =self.tmc_uart.read_int(reg.DRVSTATUS)
+        drvstatus =self.tmc_uart.read_int(tmc_reg.DRVSTATUS)
         self.log(bin(drvstatus), Loglevel.INFO.value)
-        if drvstatus & reg.stst:
+        if drvstatus & tmc_reg.stst:
             self.log("Info: motor is standing still")
         else:
             self.log("Info: motor is running")
 
-        if drvstatus & reg.stealth:
+        if drvstatus & tmc_reg.stealth:
             self.log("Info: motor is running on StealthChop")
         else:
             self.log("Info: motor is running on SpreadCycle")
 
-        cs_actual = drvstatus & reg.cs_actual
+        cs_actual = drvstatus & tmc_reg.cs_actual
         cs_actual = cs_actual >> 16
         self.log("CS actual: "+str(cs_actual))
 
-        if drvstatus & reg.olb:
+        if drvstatus & tmc_reg.olb:
             self.log("Warning: Open load detected on phase B")
 
-        if drvstatus & reg.ola:
+        if drvstatus & tmc_reg.ola:
             self.log("Warning: Open load detected on phase A")
 
-        if drvstatus & reg.s2vsb:
+        if drvstatus & tmc_reg.s2vsb:
             self.log("""Error: Short on low-side MOSFET detected on phase B.
                      The driver becomes disabled""")
 
-        if drvstatus & reg.s2vsa:
+        if drvstatus & tmc_reg.s2vsa:
             self.log("""Error: Short on low-side MOSFET detected on phase A.
                      The driver becomes disabled""")
 
-        if drvstatus & reg.s2gb:
+        if drvstatus & tmc_reg.s2gb:
             self.log("Error: Short to GND detected on phase B. The driver becomes disabled.")
 
-        if drvstatus & reg.s2ga:
+        if drvstatus & tmc_reg.s2ga:
             self.log("Error: Short to GND detected on phase A. The driver becomes disabled.")
 
-        if drvstatus & reg.ot:
+        if drvstatus & tmc_reg.ot:
             self.log("Error: Driver Overheating!")
 
-        if drvstatus & reg.otpw:
+        if drvstatus & tmc_reg.otpw:
             self.log("Warning: Driver Overheating Prewarning!")
 
         self.log("---")
@@ -286,39 +313,42 @@ class TMC_2209:
     def readGCONF(self):
         """
         read the register Adress "GCONF" and prints all current setting
+
+            Returns:
+                gconf (int): 10bit GCONF Register
         """
         self.log("---")
         self.log("GENERAL CONFIG")
-        gconf = self.tmc_uart.read_int(reg.GCONF)
+        gconf = self.tmc_uart.read_int(tmc_reg.GCONF)
         self.log(bin(gconf), Loglevel.INFO.value)
 
-        if gconf & reg.i_scale_analog:
+        if gconf & tmc_reg.i_scale_analog:
             self.log("Driver is using voltage supplied to VREF as current reference")
         else:
             self.log("Driver is using internal reference derived from 5VOUT")
-        if gconf & reg.internal_rsense:
+        if gconf & tmc_reg.internal_rsense:
             self.log("Internal sense resistors. Use current supplied into VREF as reference.")
             self.log("VREF pin internally is driven to GND in this mode.")
             self.log("This will most likely destroy your driver!!!")
             raise SystemExit
         self.log("Operation with external sense resistors")
-        if gconf & reg.en_spreadcycle:
+        if gconf & tmc_reg.en_spreadcycle:
             self.log("SpreadCycle mode enabled")
         else:
             self.log("StealthChop PWM mode enabled")
-        if gconf & reg.shaft:
+        if gconf & tmc_reg.shaft:
             self.log("Inverse motor direction")
         else:
             self.log("normal motor direction")
-        if gconf & reg.index_otpw:
+        if gconf & tmc_reg.index_otpw:
             self.log("INDEX pin outputs overtemperature prewarning flag")
         else:
             self.log("INDEX shows the first microstep position of sequencer")
-        if gconf & reg.index_step:
+        if gconf & tmc_reg.index_step:
             self.log("INDEX output shows step pulses from internal pulse generator")
         else:
             self.log("INDEX output as selected by index_otpw")
-        if gconf & reg.mstep_reg_select:
+        if gconf & tmc_reg.mstep_reg_select:
             self.log("Microstep resolution selected by MSTEP register")
         else:
             self.log("Microstep resolution selected by pins MS1, MS2")
@@ -331,17 +361,20 @@ class TMC_2209:
     def readGSTAT(self):
         """
         read the register Adress "GSTAT" and prints all current setting
+
+            Returns:
+                gstat (int): 3bit GSTAT Register
         """
         self.log("---")
         self.log("GSTAT")
-        gstat = self.tmc_uart.read_int(reg.GSTAT)
+        gstat = self.tmc_uart.read_int(tmc_reg.GSTAT)
         self.log(bin(gstat), Loglevel.INFO.value)
-        if gstat & reg.reset:
+        if gstat & tmc_reg.reset:
             self.log("The Driver has been reset since the last read access to GSTAT")
-        if gstat & reg.drv_err:
+        if gstat & tmc_reg.drv_err:
             self.log("""The driver has been shut down due to overtemperature or
                      short circuit detection since the last read access""")
-        if gstat & reg.uv_cp:
+        if gstat & tmc_reg.uv_cp:
             self.log("Undervoltage on the charge pump. The driver is disabled in this case")
         self.log("---")
         return gstat
@@ -353,39 +386,42 @@ class TMC_2209:
         read the register Adress "GSTAT" and prints all current setting
         """
         self.log("clearing GSTAT", Loglevel.INFO.value)
-        gstat = self.tmc_uart.read_int(reg.GSTAT)
+        gstat = self.tmc_uart.read_int(tmc_reg.GSTAT)
 
-        gstat = self.tmc_uart.set_bit(gstat, reg.reset)
-        gstat = self.tmc_uart.set_bit(gstat, reg.drv_err)
+        gstat = self.tmc_uart.set_bit(gstat, tmc_reg.reset)
+        gstat = self.tmc_uart.set_bit(gstat, tmc_reg.drv_err)
 
-        self.tmc_uart.write_reg_check(reg.GSTAT, gstat)
+        self.tmc_uart.write_reg_check(tmc_reg.GSTAT, gstat)
 
 
 
     def readIOIN(self):
         """
         read the register Adress "IOIN" and prints all current setting
+
+            Returns:
+                ioin (int): 10+8bit IOIN Register
         """
         self.log("---")
         self.log("INPUTS")
-        ioin = self.tmc_uart.read_int(reg.IOIN)
+        ioin = self.tmc_uart.read_int(tmc_reg.IOIN)
         self.log(bin(ioin), Loglevel.INFO.value)
-        if ioin & reg.io_spread:
+        if ioin & tmc_reg.io_spread:
             self.log("spread is high")
         else:
             self.log("spread is low")
 
-        if ioin & reg.io_dir:
+        if ioin & tmc_reg.io_dir:
             self.log("dir is high")
         else:
             self.log("dir is low")
 
-        if ioin & reg.io_step:
+        if ioin & tmc_reg.io_step:
             self.log("step is high")
         else:
             self.log("step is low")
 
-        if ioin & reg.io_enn:
+        if ioin & tmc_reg.io_enn:
             self.log("en is high")
         else:
             self.log("en is low")
@@ -398,18 +434,21 @@ class TMC_2209:
     def readCHOPCONF(self):
         """
         read the register Adress "CHOPCONF" and prints all current setting
+
+            Returns:
+                chopconf (int): 3bit CHOPCONF Register
         """
         self.log("---")
         self.log("CHOPPER CONTROL")
-        chopconf = self.tmc_uart.read_int(reg.CHOPCONF)
+        chopconf = self.tmc_uart.read_int(tmc_reg.CHOPCONF)
         self.log(bin(chopconf), Loglevel.INFO.value)
 
         self.log("native "+str(self.get_microstepping_resolution())+" microstep setting")
 
-        if chopconf & reg.intpol:
+        if chopconf & tmc_reg.intpol:
             self.log("interpolation to 256 microsteps")
 
-        if chopconf & reg.vsense:
+        if chopconf & tmc_reg.vsense:
             self.log("1: High sensitivity, low sense resistor voltage")
         else:
             self.log("0: Low sensitivity, high sense resistor voltage")
@@ -422,36 +461,49 @@ class TMC_2209:
     def set_motor_enabled(self, en):
         """
         enables or disables the motor current output
+
+            Parameters:
+                en (bool): whether the motor current output should be enabled
         """
         GPIO.output(self._pin_en, not en)
         self.log(f"Motor output active: {en}", Loglevel.INFO.value)
 
 
 
-    def do_homing(self, diag_pin, revolutions, threshold=None):
+    def do_homing(self, diag_pin, revolutions = 10, threshold = None, speed_rpm = None):
         """
-        homes the motor in the given direction using stallguard
-        1. param: DIAG pin
-        2. param: maximum number of revolutions. Can be negative for inverse direction
-        3. param(optional: StallGuard detection threshold
-        returns true when homing was successful
+        homes the motor in the given direction using stallguard.
+        this method is using vactual to move the motor and an interrupt on the DIAG pin
+
+            Parameters
+                diag_pin (int): DIAG pin number
+                revolutions (int): max number of revolutions. Can be negative for inverse direction
+                threshold (int): optional; StallGuard detection threshold
+                speed_rpm (float): optional; speed in revolutions per minute
+
+            Returns:
+                not homing_failed (bool): true when homing was successful
         """
         if threshold is not None:
             self._sg_threshold = threshold
+        if speed_rpm is None:
+            speed_rpm = tmc_math.steps_to_rps(self._max_speed_homing, self._steps_per_rev)*60
 
         self.log("---", Loglevel.INFO.value)
         self.log("homing", Loglevel.INFO.value)
 
+        # StallGuard only works with StealthChop
         self.set_spreadcycle(0)
 
         self.log("Stallguard threshold:"+str(self._sg_threshold), Loglevel.DEBUG.value)
 
-        self.set_stallguard_callback(diag_pin, self._sg_threshold, self.stop)
+        self.set_stallguard_callback(diag_pin, self._sg_threshold, self.stop,
+                                     0.5*tmc_math.rps_to_steps(speed_rpm/60, self._steps_per_rev))
 
-        homing_failed = self.set_vactual_rpm(30, revolutions=revolutions)
+        homing_failed = self.set_vactual_rpm(speed_rpm, revolutions=revolutions)
 
         if homing_failed:
-            self.log("homing failed", Loglevel.ERROR.value)
+            self.log("homing failed", Loglevel.INFO.value)
         else:
             self.log("homing successful",Loglevel.INFO.value)
 
@@ -462,12 +514,15 @@ class TMC_2209:
 
 
 
-    def do_homing2(self, direction, threshold=None):
+    def do_homing2(self, revolutions, threshold=None):
         """
         homes the motor in the given direction using stallguard
-        old function, uses STEP/DIR 
-        1. param: direction
-        2. param(optional: StallGuard detection threshold
+        old function, uses STEP/DIR to move the motor and pulls the StallGuard result
+        from the interface
+
+            Parameters
+                revolutions (int): max number of revolutions. Can be negative for inverse direction
+                threshold (int): optional; StallGuard detection threshold
         """
         sg_results = []
 
@@ -479,24 +534,25 @@ class TMC_2209:
 
         self.log("Stallguard threshold:"+str(self._sg_threshold), Loglevel.DEBUG.value)
 
-        self.set_direction_pin(direction)
+        self.set_direction_pin(revolutions > 0)
+
+        # StallGuard only works with StealthChop
         self.set_spreadcycle(0)
 
-        if direction == 1:
-            self._target_pos = self._steps_per_revolution * 10
-        else:
-            self._target_pos = -self._steps_per_revolution * 10
+        self._target_pos = self._steps_per_rev * revolutions
         self._step_interval = 0
         self._speed = 0.0
         self._n = 0
         self.set_acceleration(10000)
         self.set_max_speed(self._max_speed_homing)
+        coolstep_thres = tmc_math.steps_to_tstep(self._max_speed_homing*0.5, self._msres)
+        self.set_coolstep_threshold(coolstep_thres)
         self.compute_new_speed()
 
 
         step_counter=0
-        #self.log("Steps per Revolution: "+str(self._steps_per_revolution))
-        while step_counter<self._steps_per_revolution:
+        #self.log("Steps per Revolution: "+str(self._steps_per_rev))
+        while step_counter<self._target_pos:
             if self.run_speed(): #returns true, when a step is made
                 step_counter += 1
                 self.compute_new_speed()
@@ -507,13 +563,13 @@ class TMC_2209:
                     if sg_result_average < self._sg_threshold:
                         break
 
-        if step_counter<self._steps_per_revolution:
+        if step_counter<self._steps_per_rev:
             self.log("homing successful",Loglevel.INFO.value)
             self.log("Stepcounter: "+str(step_counter),Loglevel.DEBUG.value)
             self.log(str(sg_results),Loglevel.DEBUG.value)
             self._current_pos = 0
         else:
-            self.log("homing failed", Loglevel.ERROR.value)
+            self.log("homing failed", Loglevel.INFO.value)
             self.log("Stepcounter: "+str(step_counter), Loglevel.DEBUG.value)
             self.log(str(sg_results),Loglevel.DEBUG.value)
 
@@ -524,6 +580,9 @@ class TMC_2209:
     def get_current_position(self):
         """
         returns the current motor position in microsteps
+
+            Returns:
+                current_pos (bool): current motor position
         """
         return self._current_pos
 
@@ -532,6 +591,9 @@ class TMC_2209:
     def set_current_position(self, new_pos):
         """
         overwrites the current motor position in microsteps
+
+            Parameters:
+                new_pos (bool): new position
         """
         self._current_pos = new_pos
 
@@ -549,6 +611,9 @@ class TMC_2209:
     def set_direction_pin(self, direction):
         """
         sets the motor shaft direction to the given value: 0 = CCW; 1 = CW
+
+            Parameters:
+                direction (bool): motor shaft direction: 0 = CCW; 1 = CW
         """
         self._direction = direction
         GPIO.output(self._pin_dir, direction)
@@ -558,48 +623,60 @@ class TMC_2209:
     def get_direction_reg(self):
         """
         returns the motor shaft direction: 0 = CCW; 1 = CW
+
+            Returns:
+                direction (bool): motor shaft direction: 0 = CCW; 1 = CW
         """
-        gconf = self.tmc_uart.read_int(reg.GCONF)
-        return gconf & reg.shaft
+        gconf = self.tmc_uart.read_int(tmc_reg.GCONF)
+        return gconf & tmc_reg.shaft
 
 
 
     def set_direction_reg(self, direction):
         """
         sets the motor shaft direction to the given value: 0 = CCW; 1 = CW
+
+            Parameters:
+                direction (bool): motor shaft direction: 0 = CCW; 1 = CW
         """
-        gconf = self.tmc_uart.read_int(reg.GCONF)
+        gconf = self.tmc_uart.read_int(tmc_reg.GCONF)
         if direction:
             self.log("write inverse motor direction", Loglevel.INFO.value)
-            gconf = self.tmc_uart.set_bit(gconf, reg.shaft)
+            gconf = self.tmc_uart.set_bit(gconf, tmc_reg.shaft)
         else:
             self.log("write normal motor direction", Loglevel.INFO.value)
-            gconf = self.tmc_uart.clear_bit(gconf, reg.shaft)
-        self.tmc_uart.write_reg_check(reg.GCONF, gconf)
+            gconf = self.tmc_uart.clear_bit(gconf, tmc_reg.shaft)
+        self.tmc_uart.write_reg_check(tmc_reg.GCONF, gconf)
 
 
 
     def get_iscale_analog(self):
         """
         return whether Vref (1) or 5V (0) is used for current scale
+
+            Returns:
+                en (bool): whether Vref (1) or 5V (0) is used for current scale
         """
-        gconf = self.tmc_uart.read_int(reg.GCONF)
-        return gconf & reg.i_scale_analog
+        gconf = self.tmc_uart.read_int(tmc_reg.GCONF)
+        return gconf & tmc_reg.i_scale_analog
 
 
 
     def set_iscale_analog(self,en):
         """
         sets Vref (1) or 5V (0) for current scale
+
+            Parameters:
+                en (bool): whether Vref (1) or 5V (0) is used for current scale
         """
-        gconf = self.tmc_uart.read_int(reg.GCONF)
+        gconf = self.tmc_uart.read_int(tmc_reg.GCONF)
         if en:
             self.log("activated Vref for current scale", Loglevel.INFO.value)
-            gconf = self.tmc_uart.set_bit(gconf, reg.i_scale_analog)
+            gconf = self.tmc_uart.set_bit(gconf, tmc_reg.i_scale_analog)
         else:
             self.log("activated 5V-out for current scale", Loglevel.INFO.value)
-            gconf = self.tmc_uart.clear_bit(gconf, reg.i_scale_analog)
-        self.tmc_uart.write_reg_check(reg.GCONF, gconf)
+            gconf = self.tmc_uart.clear_bit(gconf, tmc_reg.i_scale_analog)
+        self.tmc_uart.write_reg_check(tmc_reg.GCONF, gconf)
 
 
 
@@ -608,9 +685,12 @@ class TMC_2209:
         returns which sense resistor voltage is used for current scaling
         0: Low sensitivity, high sense resistor voltage
         1: High sensitivity, low sense resistor voltage
+
+            Returns:
+                en (bool): whether high sensitivity should is used
         """
-        chopconf = self.tmc_uart.read_int(reg.CHOPCONF)
-        return chopconf & reg.vsense
+        chopconf = self.tmc_uart.read_int(tmc_reg.CHOPCONF)
+        return chopconf & tmc_reg.vsense
 
 
 
@@ -619,64 +699,81 @@ class TMC_2209:
         sets which sense resistor voltage is used for current scaling
         0: Low sensitivity, high sense resistor voltage
         1: High sensitivity, low sense resistor voltage
+
+            Parameters:
+                en (bool): whether high sensitivity should be used
         """
-        chopconf = self.tmc_uart.read_int(reg.CHOPCONF)
+        chopconf = self.tmc_uart.read_int(tmc_reg.CHOPCONF)
         if en:
             self.log("activated High sensitivity, low sense resistor voltage", Loglevel.INFO.value)
-            chopconf = self.tmc_uart.set_bit(chopconf, reg.vsense)
+            chopconf = self.tmc_uart.set_bit(chopconf, tmc_reg.vsense)
         else:
             self.log("activated Low sensitivity, high sense resistor voltage", Loglevel.INFO.value)
-            chopconf = self.tmc_uart.clear_bit(chopconf, reg.vsense)
-        self.tmc_uart.write_reg_check(reg.CHOPCONF, chopconf)
+            chopconf = self.tmc_uart.clear_bit(chopconf, tmc_reg.vsense)
+        self.tmc_uart.write_reg_check(tmc_reg.CHOPCONF, chopconf)
 
 
 
     def get_internal_rsense(self):
         """
         returns which sense resistor voltage is used for current scaling
-        0: Low sensitivity, high sense resistor voltage
-        1: High sensitivity, low sense resistor voltage
+        0: Operation with external sense resistors
+        1: Internal sense resistors. Use current supplied into
+        VREF as reference for internal sense resistor. VREF
+        pin internally is driven to GND in this mode.
+
+            Returns:
+                en (bool): which sense resistor voltage is used
         """
-        gconf = self.tmc_uart.read_int(reg.GCONF)
-        return gconf & reg.internal_rsense
+        gconf = self.tmc_uart.read_int(tmc_reg.GCONF)
+        return gconf & tmc_reg.internal_rsense
 
 
 
     def set_internal_rsense(self,en):
         """
         sets which sense resistor voltage is used for current scaling
-        0: Low sensitivity, high sense resistor voltage
-        1: High sensitivity, low sense resistor voltage
+        0: Operation with external sense resistors
+        1: Internal sense resistors. Use current supplied into
+        VREF as reference for internal sense resistor. VREF
+        pin internally is driven to GND in this mode.
+
+            Parameters:
+                en (bool):which sense resistor voltage is used; true will propably destroy your tmc
         """
-        gconf = self.tmc_uart.read_int(reg.GCONF)
+        gconf = self.tmc_uart.read_int(tmc_reg.GCONF)
         if en:
             self.log("activated internal sense resistors.", Loglevel.INFO.value)
             self.log("VREF pin internally is driven to GND in this mode.", Loglevel.INFO.value)
             self.log("This will most likely destroy your driver!!!", Loglevel.INFO.value)
             raise SystemExit
-            # gconf = self.tmc_uart.set_bit(gconf, reg.internal_rsense)
+            # gconf = self.tmc_uart.set_bit(gconf, tmc_reg.internal_rsense)
         self.log("activated operation with external sense resistors", Loglevel.INFO.value)
-        gconf = self.tmc_uart.clear_bit(gconf, reg.internal_rsense)
-        self.tmc_uart.write_reg_check(reg.GCONF, gconf)
+        gconf = self.tmc_uart.clear_bit(gconf, tmc_reg.internal_rsense)
+        self.tmc_uart.write_reg_check(tmc_reg.GCONF, gconf)
 
 
 
-    def set_irun_ihold(self, IHold, IRun, ihold_delay):
+    def set_irun_ihold(self, ihold, irun, ihold_delay):
         """
         sets the current scale (CS) for Running and Holding
         and the delay, when to be switched to Holding current
-        IHold = 0-31; IRun = 0-31; ihold_delay = 0-15
+
+            Parameters:
+                ihold (int): 0-31
+                irun (int): 0-31
+                ihold_delay (int): 0-15
         """
         ihold_irun = 0
 
-        ihold_irun = ihold_irun | IHold << 0
-        ihold_irun = ihold_irun | IRun << 8
+        ihold_irun = ihold_irun | ihold << 0
+        ihold_irun = ihold_irun | irun << 8
         ihold_irun = ihold_irun | ihold_delay << 16
         self.log("ihold_irun", Loglevel.INFO.value)
         self.log(str(bin(ihold_irun)), Loglevel.INFO.value)
 
         self.log("writing ihold_irun", Loglevel.INFO.value)
-        self.tmc_uart.write_reg_check(reg.IHOLD_IRUN, ihold_irun)
+        self.tmc_uart.write_reg_check(tmc_reg.IHOLD_IRUN, ihold_irun)
 
 
 
@@ -686,24 +783,33 @@ class TMC_2209:
         0: PDN_UART controls standstill current reduction
         1: PDN_UART input function disabled. Set this bit,
         when using the UART interface!
+
+            Parameters:
+                pdn_disable (bool): whether pdn should be disabled
         """
-        gconf = self.tmc_uart.read_int(reg.GCONF)
+        gconf = self.tmc_uart.read_int(tmc_reg.GCONF)
         if pdn_disable:
             self.log("enabled PDN_UART", Loglevel.INFO.value)
-            gconf = self.tmc_uart.set_bit(gconf, reg.pdn_disable)
+            gconf = self.tmc_uart.set_bit(gconf, tmc_reg.pdn_disable)
         else:
             self.log("disabled PDN_UART", Loglevel.INFO.value)
-            gconf = self.tmc_uart.clear_bit(gconf, reg.pdn_disable)
-        self.tmc_uart.write_reg_check(reg.GCONF, gconf)
+            gconf = self.tmc_uart.clear_bit(gconf, tmc_reg.pdn_disable)
+        self.tmc_uart.write_reg_check(tmc_reg.GCONF, gconf)
 
 
 
     def set_current(self, run_current, hold_current_multiplier = 0.5, hold_current_delay = 10,
-                    use_vref=False, Vref = 1.2, pdn_disable = True):
+                    use_vref=False, vref = 1.2, pdn_disable = True):
         """
-        sets the current flow for the motor
-        run_current in mA
-        check whether Vref is actually 1.2V
+        sets the current flow for the motor. 
+
+            Parameters:
+                run_current (int): current during movement in mA
+                hold_current_multiplier (int): current multiplier during standstill
+                hold_current_delay (int): delay after standstill after which the current drops
+                use_vref (int): most driver have a trimpotentiometer to set the vref. optional
+                Vref (int): if you want to use Vref, you need to set the Vref Voltage
+                pdn_disable (bool): should be disabled if UART is used
         """
         cs_irun = 0
         rsense = 0.11
@@ -712,16 +818,16 @@ class TMC_2209:
         if use_vref:
             self.set_iscale_analog(True)
         else:
-            Vref=2.5
+            vref=2.5
             self.set_iscale_analog(False)
 
-        vfs = 0.325 * Vref / 2.5
+        vfs = 0.325 * vref / 2.5
         cs_irun = 32.0*1.41421*run_current/1000.0*(rsense+0.02)/vfs - 1
 
         # If Current Scale is too low, turn on high sensitivity VSsense and calculate again
         if cs_irun < 16:
             self.log("CS too low; switching to VSense True", Loglevel.INFO.value)
-            vfs = 0.180 * Vref / 2.5
+            vfs = 0.180 * vref / 2.5
             cs_irun = 32.0*1.41421*run_current/1000.0*(rsense+0.02)/vfs - 1
             self.setvsense(True)
         else: # If CS >= 16, turn off high_senser
@@ -754,59 +860,75 @@ class TMC_2209:
     def get_spreadcycle(self):
         """
         return whether spreadcycle (1) is active or stealthchop (0)
+
+            Parameters:
+                en_spread (bool): true = spreadcycle; false = stealthchop
         """
-        gconf = self.tmc_uart.read_int(reg.GCONF)
-        return gconf & reg.en_spreadcycle
+        gconf = self.tmc_uart.read_int(tmc_reg.GCONF)
+        return gconf & tmc_reg.en_spreadcycle
 
 
 
     def set_spreadcycle(self,en_spread):
         """
         enables spreadcycle (1) or stealthchop (0)
+
+            Parameters:
+                en_spread (bool): true to enable spreadcycle; false to enable stealthchop
         """
-        gconf = self.tmc_uart.read_int(reg.GCONF)
+        gconf = self.tmc_uart.read_int(tmc_reg.GCONF)
         if en_spread:
             self.log("activated Spreadcycle", Loglevel.INFO.value)
-            gconf = self.tmc_uart.set_bit(gconf, reg.en_spreadcycle)
+            gconf = self.tmc_uart.set_bit(gconf, tmc_reg.en_spreadcycle)
         else:
             self.log("activated Stealthchop", Loglevel.INFO.value)
-            gconf = self.tmc_uart.clear_bit(gconf, reg.en_spreadcycle)
-        self.tmc_uart.write_reg_check(reg.GCONF, gconf)
+            gconf = self.tmc_uart.clear_bit(gconf, tmc_reg.en_spreadcycle)
+        self.tmc_uart.write_reg_check(tmc_reg.GCONF, gconf)
 
 
 
     def get_interpolation(self):
         """
         return whether the tmc inbuilt interpolation is active
+
+            Returns:
+                en (bool): true if internal µstep interpolation is enabled
         """
-        chopconf = self.tmc_uart.read_int(reg.CHOPCONF)
-        return bool(chopconf & reg.intpol)
+        chopconf = self.tmc_uart.read_int(tmc_reg.CHOPCONF)
+        return bool(chopconf & tmc_reg.intpol)
 
 
 
     def set_interpolation(self, en):
         """
         enables the tmc inbuilt interpolation of the steps to 256 microsteps
+
+            Parameters:
+                en (bool): true to enable internal µstep interpolation
         """
-        chopconf = self.tmc_uart.read_int(reg.CHOPCONF)
+        chopconf = self.tmc_uart.read_int(tmc_reg.CHOPCONF)
 
         if en:
-            chopconf = self.tmc_uart.set_bit(chopconf, reg.intpol)
+            chopconf = self.tmc_uart.set_bit(chopconf, tmc_reg.intpol)
         else:
-            chopconf = self.tmc_uart.clear_bit(chopconf, reg.intpol)
+            chopconf = self.tmc_uart.clear_bit(chopconf, tmc_reg.intpol)
 
         self.log("writing microstep interpolation setting: "+str(en), Loglevel.INFO.value)
-        self.tmc_uart.write_reg_check(reg.CHOPCONF, chopconf)
+        self.tmc_uart.write_reg_check(tmc_reg.CHOPCONF, chopconf)
 
 
 
     def get_microstepping_resolution(self):
         """
         returns the current native microstep resolution (1-256)
-        """
-        chopconf = self.tmc_uart.read_int(reg.CHOPCONF)
 
-        msresdezimal = chopconf & (reg.msres0 | reg.msres1 | reg.msres2 | reg.msres3)
+            Returns:
+                msres (int): µstep resolution
+        """
+        chopconf = self.tmc_uart.read_int(tmc_reg.CHOPCONF)
+
+        msresdezimal = chopconf & (tmc_reg.msres0 | tmc_reg.msres1 |
+                                   tmc_reg.msres2 | tmc_reg.msres3)
 
         msresdezimal = msresdezimal >> 24
         msresdezimal = 8 - msresdezimal
@@ -820,21 +942,25 @@ class TMC_2209:
     def set_microstepping_resolution(self, msres):
         """
         sets the current native microstep resolution (1,2,4,8,16,32,64,128,256)
+
+            Parameters:
+                msres (int): µstep resolution; has to be a power of 2 or 1 for fullstep
         """
-        chopconf = self.tmc_uart.read_int(reg.CHOPCONF)
+        chopconf = self.tmc_uart.read_int(tmc_reg.CHOPCONF)
         #setting all bits to zero
-        chopconf = chopconf & (~reg.msres0 | ~reg.msres1 | ~reg.msres2 | ~reg.msres3)
+        chopconf = chopconf & (~tmc_reg.msres0 | ~tmc_reg.msres1 |
+                               ~tmc_reg.msres2 | ~tmc_reg.msres3)
         msresdezimal = int(math.log(msres, 2))
         msresdezimal = 8 - msresdezimal
         chopconf = int(chopconf) & int(4043309055)
         chopconf = chopconf | msresdezimal <<24
 
         self.log("writing "+str(msres)+" microstep setting", Loglevel.INFO.value)
-        self.tmc_uart.write_reg_check(reg.CHOPCONF, chopconf)
+        self.tmc_uart.write_reg_check(tmc_reg.CHOPCONF, chopconf)
 
         self.set_mstep_resolution_reg_select(True)
 
-        self.read_steps_per_revolution()
+        self.read_steps_per_rev()
 
         return True
 
@@ -845,33 +971,44 @@ class TMC_2209:
         sets the register bit "mstep_reg_select" to 1 or 0 depending to the given value.
         this is needed to set the microstep resolution via UART
         this method is called by "set_microstepping_resolution"
+
+            Parameters:
+                en (bool): true to set µstep resolution via UART
         """
-        gconf = self.tmc_uart.read_int(reg.GCONF)
+        gconf = self.tmc_uart.read_int(tmc_reg.GCONF)
 
         if en is True:
-            gconf = self.tmc_uart.set_bit(gconf, reg.mstep_reg_select)
+            gconf = self.tmc_uart.set_bit(gconf, tmc_reg.mstep_reg_select)
         else:
-            gconf = self.tmc_uart.clear_bit(gconf, reg.mstep_reg_select)
+            gconf = self.tmc_uart.clear_bit(gconf, tmc_reg.mstep_reg_select)
 
         self.log("writing MStep Reg Select: "+str(en), Loglevel.INFO.value)
-        self.tmc_uart.write_reg_check(reg.GCONF, gconf)
+        self.tmc_uart.write_reg_check(tmc_reg.GCONF, gconf)
 
 
 
-    def read_steps_per_revolution(self):
+    def read_steps_per_rev(self):
         """
-        returns how many steps are needed for one revolution
+        returns how many steps are needed for one revolution.
+        this reads the value from the tmc driver.
+
+            Returns:
+                steps_per_rev (int): Steps per revolution
         """
-        self._steps_per_revolution = 200*self.get_microstepping_resolution()
-        return self._steps_per_revolution
+        self._steps_per_rev = self._fullsteps_per_rev*self.get_microstepping_resolution()
+        return self._steps_per_rev
 
 
 
-    def get_steps_per_revolution(self):
+    def get_steps_per_rev(self):
         """
-        returns how many steps are needed for one revolution
+        returns how many steps are needed for one revolution.
+        this gets the cached value from the library.
+
+            Returns:
+                steps_per_rev (int): Steps per revolution
         """
-        return self._steps_per_revolution
+        return self._steps_per_rev
 
 
 
@@ -880,8 +1017,11 @@ class TMC_2209:
         reads the interface transmission counter from the tmc register
         this value is increased on every succesfull write access
         can be used to verify a write access
+
+            Returns:
+                ifcnt (int): 8bit IFCNT Register
         """
-        ifcnt = self.tmc_uart.read_int(reg.IFCNT)
+        ifcnt = self.tmc_uart.read_int(tmc_reg.IFCNT)
         self.log("Interface Transmission Counter: "+str(ifcnt), Loglevel.INFO.value)
         return ifcnt
 
@@ -889,11 +1029,12 @@ class TMC_2209:
 
     def get_tstep(self):
         """
-        return the current stallguard result
-        its will be calculated with every fullstep
-        higher values means a lower motor load
+        reads the current tstep from the driver register
+
+            Returns:
+                tstep (int): TStep time
         """
-        tstep = self.tmc_uart.read_int(reg.TSTEP)
+        tstep = self.tmc_uart.read_int(tmc_reg.TSTEP)
         return tstep
 
 
@@ -905,6 +1046,16 @@ class TMC_2209:
         VACTUAL allows moving the motor by UART control.
         It gives the motor velocity in +-(2^23)-1 [μsteps / t]
         0: Normal operation. Driver reacts to STEP input
+
+            Parameters:
+                vactual (int): value for vactual
+                duration (int): after this vactual will be set to 0
+                acceleration (int): use this for a velocity ramp
+                show_stallguard_result (bool): prints StallGuard Result during movement
+                show_tstep (bool): prints TStep during movement
+
+            Returns:
+                stop (enum): how the movement was finished
         """
         self._stop = StopMode.NO
         current_vactual = 0
@@ -920,7 +1071,7 @@ class TMC_2209:
 
         self.log("writing vactual", Loglevel.INFO.value)
         if acceleration == 0:
-            self.tmc_uart.write_reg_check(reg.VACTUAL, vactual)
+            self.tmc_uart.write_reg_check(tmc_reg.VACTUAL, vactual)
 
         if duration == 0:
             return -1
@@ -936,11 +1087,11 @@ class TMC_2209:
                     time_to_stop = current_time-1
             if acceleration != 0 and current_time > time_to_stop:
                 current_vactual -= acceleration*sleeptime
-                self.tmc_uart.write_reg_check(reg.VACTUAL, int(round(current_vactual)))
+                self.tmc_uart.write_reg_check(tmc_reg.VACTUAL, int(round(current_vactual)))
                 time.sleep(sleeptime)
             elif acceleration != 0 and abs(current_vactual)<abs(vactual):
                 current_vactual += acceleration*sleeptime
-                self.tmc_uart.write_reg_check(reg.VACTUAL, int(round(current_vactual)))
+                self.tmc_uart.write_reg_check(tmc_reg.VACTUAL, int(round(current_vactual)))
                 time.sleep(sleeptime)
             if show_stallguard_result:
                 self.log("StallGuard result: "+str(self.get_stallguard_result()),
@@ -950,7 +1101,7 @@ class TMC_2209:
                 self.log("TStep result: "+str(self.get_tstep()), Loglevel.INFO.value)
                 time.sleep(0.1)
             current_time = time.time()
-        self.tmc_uart.write_reg_check(reg.VACTUAL, 0)
+        self.tmc_uart.write_reg_check(tmc_reg.VACTUAL, 0)
         return self._stop
 
 
@@ -961,13 +1112,24 @@ class TMC_2209:
         rotation speed in revolutions per second
         With internal oscillator:
         VACTUAL[2209] = v[Hz] / 0.715Hz
+
+            Parameters:
+                rps (int): value for vactual in rps
+                duration (int): after this vactual will be set to 0
+                revolutions (int): after this vactual will be set to 0
+                acceleration (int): use this for a velocity ramp
+
+            Returns:
+                stop (enum): how the movement was finished
         """
-        vactual = rps/0.715*self._steps_per_revolution
-        if revolutions!=0:
-            duration = abs(revolutions/rps)
+        if revolutions == 0:
+            return -1
+        print("vactual")
+        vactual = tmc_math.rps_to_vactual(rps, self._steps_per_rev)
+        duration = abs(revolutions/rps)
         if revolutions<0:
             vactual = -vactual
-        return self.set_vactual(int(round(vactual)), duration, acceleration=acceleration)
+        return self.set_vactual(vactual, duration, acceleration=acceleration)
 
 
 
@@ -975,6 +1137,15 @@ class TMC_2209:
         """
         converts the rps parameter to a vactual value which represents
         rotation speed in revolutions per minute
+
+            Parameters:
+                rpm (int): value for vactual in rpm
+                duration (int): after this vactual will be set to 0
+                revolutions (int): after this vactual will be set to 0
+                acceleration (int): use this for a velocity ramp
+
+            Returns:
+                stop (enum): how the movement was finished
         """
         return self.set_vactual_rps(rpm/60, duration, revolutions, acceleration)
 
@@ -989,7 +1160,7 @@ class TMC_2209:
             Returns:
                 sg_result (int): StallGuard Result
         """
-        sg_result = self.tmc_uart.read_int(reg.SG_RESULT)
+        sg_result = self.tmc_uart.read_int(tmc_reg.SG_RESULT)
         return sg_result
 
 
@@ -1009,7 +1180,7 @@ class TMC_2209:
         self.log(str(bin(threshold)), Loglevel.INFO.value)
 
         self.log("writing sgthrs", Loglevel.INFO.value)
-        self.tmc_uart.write_reg_check(reg.SGTHRS, threshold)
+        self.tmc_uart.write_reg_check(tmc_reg.SGTHRS, threshold)
 
 
 
@@ -1017,13 +1188,16 @@ class TMC_2209:
         """
         This  is  the  lower  threshold  velocity  for  switching  
         on  smart energy CoolStep and StallGuard to DIAG output. (unsigned)
+
+            Parameters:
+                threshold (int): threshold  velocity
         """
 
         self.log("tcoolthrs", Loglevel.INFO.value)
         self.log(str(bin(threshold)), Loglevel.INFO.value)
 
         self.log("writing tcoolthrs", Loglevel.INFO.value)
-        self.tmc_uart.write_reg_check(reg.TCOOLTHRS, threshold)
+        self.tmc_uart.write_reg_check(tmc_reg.TCOOLTHRS, threshold)
 
 
 
@@ -1045,7 +1219,7 @@ class TMC_2209:
                  Loglevel.INFO.value)
 
         self.set_stallguard_threshold(threshold)
-        self.set_coolstep_threshold(int(round(12000000 / (min_speed * 256 / self._msres))))
+        self.set_coolstep_threshold(tmc_math.steps_to_tstep(min_speed, self._msres))
         self._sg_delay = ignore_delay
         self._sg_callback = callback
         self._pin_stallguard = pin_stallguard
@@ -1057,18 +1231,22 @@ class TMC_2209:
 
 
 
-    def stallguard_callback(self, channel):
+    def stallguard_callback(self, gpio_pin):
         """
         the callback function for StallGuard.
         only checks whether the duration of the current movement is longer than
         _sg_delay and then calls the actual callback
+
+            Parameters:
+                gpio_pin (int): pin number of the interrupt pin
         """
+        del gpio_pin
         if self._sg_callback is None:
             self.log("StallGuard callback is None", Loglevel.DEBUG.value)
             return
         if time.time()<=self._starttime+self._sg_delay and self._sg_delay != 0:
             return
-        self._sg_callback(channel)
+        self._sg_callback()
 
 
 
@@ -1076,8 +1254,11 @@ class TMC_2209:
         """
         returns the current Microstep counter.
         Indicates actual position in the microstep table for CUR_A
+
+            Returns:
+                mscnt (int): current Microstep counter
         """
-        mscnt = self.tmc_uart.read_int(reg.MSCNT)
+        mscnt = self.tmc_uart.read_int(tmc_reg.MSCNT)
         return mscnt
 
 
@@ -1086,6 +1267,12 @@ class TMC_2209:
         """
         returns the current Microstep counter.
         Indicates actual position in the microstep table for CUR_A
+
+            Parameters:
+                offset (int): offset added to the current step counter
+
+            Returns:
+                step (int): current Microstep counter convertet to steps
         """
         step = (self.get_microstep_counter()-64)*(self._msres*4)/1024
         step = (4*self._msres)-step-1
@@ -1097,6 +1284,9 @@ class TMC_2209:
     def set_max_speed(self, speed):
         """
         sets the maximum motor speed in µsteps per second
+
+            Parameters:
+                speed (int): maximum speed in microsteps/sec
         """
         if speed < 0.0:
             speed = -speed
@@ -1113,6 +1303,9 @@ class TMC_2209:
     def set_max_speed_fullstep(self, speed):
         """
         sets the maximum motor speed in fullsteps per second
+
+            Parameters:
+                speed (int): maximum speed in fullsteps/sec
         """
         self.set_max_speed(speed*self._msres)
 
@@ -1121,6 +1314,9 @@ class TMC_2209:
     def get_max_speed(self):
         """
         returns the maximum motor speed in steps per second
+
+            Returns:
+                max_speed (int): current maximum speed in steps/sec
         """
         return self._max_speed
 
@@ -1129,6 +1325,9 @@ class TMC_2209:
     def set_acceleration(self, acceleration):
         """
         sets the motor acceleration/decceleration in µsteps per sec per sec
+
+            Parameters:
+                acceleration (int): acceleration/decceleration in µsteps per sec per sec
         """
         if acceleration == 0.0:
             return
@@ -1146,6 +1345,9 @@ class TMC_2209:
     def set_acceleration_fullstep(self, acceleration):
         """
         sets the motor acceleration/decceleration in fullsteps per sec per sec
+
+            Parameters:
+                acceleration (int): acceleration/decceleration in fullsteps per sec per sec
         """
         self.set_acceleration(acceleration*self._msres)
 
@@ -1154,6 +1356,9 @@ class TMC_2209:
     def get_acceleration(self):
         """
         returns the motor acceleration/decceleration in steps per sec per sec
+
+            Returns:
+                acceleration (int): acceleration/decceleration in µsteps per sec per sec
         """
         return self._acceleration
 
@@ -1162,6 +1367,9 @@ class TMC_2209:
     def stop(self, stop_mode = StopMode.HARDSTOP):
         """
         stop the current movement
+
+            Parameters:
+                stop_mode (enum): whether the movement should be stopped immediatly or softly
         """
         self._stop = stop_mode
 
@@ -1170,6 +1378,9 @@ class TMC_2209:
     def get_movement_phase(self):
         """
         return the current Movement Phase
+
+            Returns:
+                movement_phase (enum): current Movement Phase
         """
         return self._movement_phase
 
@@ -1182,6 +1393,13 @@ class TMC_2209:
         blocks the code until finished or stopped from a different thread!
         returns true when the movement if finshed normally and false,
         when the movement was stopped
+
+            Parameters:
+                steps (int): amount of steps; can be negative
+                movement_abs_rel (enum): whether the movement should be absolut or relative
+
+            Returns:
+                stop (enum): how the movement was finished
         """
         if movement_abs_rel is None:
             movement_abs_rel = self._movement_abs_rel
@@ -1210,8 +1428,15 @@ class TMC_2209:
         runs the motor to the given position.
         with acceleration and deceleration
         blocks the code until finished!
+
+            Parameters:
+                revolutions (int): amount of revs; can be negative
+                movement_abs_rel (enum): whether the movement should be absolut or relative
+
+            Returns:
+                stop (enum): how the movement was finished
         """
-        return self.run_to_position_steps(round(revolutions * self._steps_per_revolution),
+        return self.run_to_position_steps(round(revolutions * self._steps_per_rev),
                                           movement_absolute_relative)
 
 
@@ -1223,6 +1448,13 @@ class TMC_2209:
         does not block the code
         returns true when the movement if finshed normally and false,
         when the movement was stopped
+
+            Parameters:
+                steps (int): amount of steps; can be negative
+                movement_abs_rel (enum): whether the movement should be absolut or relative
+
+            Returns:
+                stop (enum): how the movement was finished
         """
         self._movement_thread = threading.Thread(target=self.run_to_position_steps,
                                                  args=(steps, movement_abs_rel))
@@ -1235,8 +1467,15 @@ class TMC_2209:
         runs the motor to the given position.
         with acceleration and deceleration
         does not block the code
+
+            Parameters:
+                revolutions (int): amount of revs; can be negative
+                movement_abs_rel (enum): whether the movement should be absolut or relative
+
+            Returns:
+                stop (enum): how the movement was finished
         """
-        return self.run_to_position_steps_threaded(round(revolutions * self._steps_per_revolution),
+        return self.run_to_position_steps_threaded(round(revolutions * self._steps_per_rev),
                                                    movement_absolute_relative)
 
 
@@ -1247,6 +1486,9 @@ class TMC_2209:
         if startet threaded
         returns true when the movement if finshed normally and false,
         when the movement was stopped
+
+            Returns:
+                stop (enum): how the movement was finished
         """
         self._movement_thread.join()
         return self._stop
@@ -1405,11 +1647,11 @@ class TMC_2209:
         GPIO.output(self._pin_en, GPIO.HIGH)
         time.sleep(0.1)
         ioin = self.readIOIN()
-        if not ioin & reg.io_dir:
+        if not ioin & tmc_reg.io_dir:
             pin_dir_ok = False
-        if not ioin & reg.io_step:
+        if not ioin & tmc_reg.io_step:
             pin_step_ok = False
-        if not ioin & reg.io_enn:
+        if not ioin & tmc_reg.io_enn:
             pin_en_ok = False
 
         GPIO.output(self._pin_step, GPIO.LOW)
@@ -1417,11 +1659,11 @@ class TMC_2209:
         GPIO.output(self._pin_en, GPIO.LOW)
         time.sleep(0.1)
         ioin = self.readIOIN()
-        if ioin & reg.io_dir:
+        if ioin & tmc_reg.io_dir:
             pin_dir_ok = False
-        if ioin & reg.io_step:
+        if ioin & tmc_reg.io_step:
             pin_step_ok = False
-        if ioin & reg.io_enn:
+        if ioin & tmc_reg.io_enn:
             pin_en_ok = False
 
         GPIO.output(self._pin_step, GPIO.HIGH)
@@ -1429,11 +1671,11 @@ class TMC_2209:
         GPIO.output(self._pin_en, GPIO.HIGH)
         time.sleep(0.1)
         ioin = self.readIOIN()
-        if not ioin & reg.io_dir:
+        if not ioin & tmc_reg.io_dir:
             pin_dir_ok = False
-        if not ioin & reg.io_step:
+        if not ioin & tmc_reg.io_step:
             pin_step_ok = False
-        if not ioin & reg.io_enn:
+        if not ioin & tmc_reg.io_enn:
             pin_en_ok = False
 
         self.set_motor_enabled(False)
@@ -1476,7 +1718,7 @@ class TMC_2209:
         """
         self.log("---")
         self.log("TEST UART")
-        result = self.tmc_uart.test_uart(reg.IOIN)
+        result = self.tmc_uart.test_uart(tmc_reg.IOIN)
 
         snd = result[0]
         rtn = result[1]
@@ -1522,6 +1764,9 @@ class TMC_2209:
         test method for tuning stallguard threshold
         run this function with your motor settings and your motor load
         the function will determine the minimum stallguard results for each movement phase
+
+            Parameters:
+                steps (int): amount of steps; can be negative
         """
 
         self.log("---", Loglevel.INFO.value)